// prismicio-types.d.ts
type Simplify<T> = {
  [KeyType in keyof T]: T[KeyType];
};
/** Content for product documents */
interface ProductDocumentData {
  /**
   * Product name field in *product*
   *
   * - **Field Type**: Title
   * - **Placeholder**: ex: Estátua Filósofo Grego
   * - **API ID Path**: product.product_name
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  product_name: prismic.TitleField;
  /**
   * images field in *product*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: product.images[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/group
   *
   */
  images: prismic.GroupField<Simplify<ProductDocumentDataImagesItem>>;
  /**
   * colection field in *product*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: ex: "Metal", "Medieval", "Estátua", "Jogos", etc...
   * - **API ID Path**: product.colection
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  colection: prismic.RichTextField;
  /**
   * description field in *product*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: dica: dê os detalhes ao chat gpt e peça ele uma descrição de um produto com os determinados detalhes
   * - **API ID Path**: product.description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  description: prismic.RichTextField;
  /**
   * initial price field in *product*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Aqui deve ser posto o preço em string, por exemplo: 149,99
   * - **API ID Path**: product.initial_price
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  initial_price: prismic.RichTextField;
  /**
   * quote field in *product*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Aqui deve ir uma citação do personagem. Pegue do google uma frase curta, é interessante!
   * - **API ID Path**: product.quote
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  quote: prismic.RichTextField;
  /**
   * dimensions field in *product*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: product.dimensions[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/group
   *
   */
  dimensions: prismic.GroupField<Simplify<ProductDocumentDataDimensionsItem>>;
  /**
   * brand field in *product*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: apenas UMA dentre: marvel, dc , disney, anime, starWars, games, statueDrawing, drawing, actionFigure
   * - **API ID Path**: product.brand
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  brand: prismic.RichTextField;
  /**
   * type field in *product*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: informe 1 entre as opções: 3D ou AF
   * - **API ID Path**: product.type
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  type: prismic.RichTextField;
  /**
   * weight field in *product*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: informe o peso. Ex: 10
   * - **API ID Path**: product.weight
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  weight: prismic.RichTextField;
}
/**
 * Item in product → images
 *
 */
export interface ProductDocumentDataImagesItem {
  /**
   * image field in *product → images*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: product.images[].image
   * - **Documentation**: https://prismic.io/docs/core-concepts/image
   *
   */
  image: prismic.ImageField<never>;
  /**
   * alt field in *product → images*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: product.images[].alt
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  alt: prismic.RichTextField;
}
/**
 * Item in product → dimensions
 *
 */
export interface ProductDocumentDataDimensionsItem {
  /**
   * sCepOrigem field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Coloque o SEU cep aqui!
   * - **API ID Path**: product.dimensions[].sceporigem
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  sceporigem: prismic.RichTextField;
  /**
   * sCepDestino field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: deixe-o assim: ""
   * - **API ID Path**: product.dimensions[].scepdestino
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  scepdestino: prismic.RichTextField;
  /**
   * nVlPeso field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Peso da encomenda, incluindo sua embalagem. Informe em quilogramas. Envelope: máximo de 1kg
   * - **API ID Path**: product.dimensions[].nvlpeso
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  nvlpeso: prismic.RichTextField;
  /**
   * nCdFormato field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Formato ex: 1 = caixa/pacote; 2 = Formato rolo/prisma; 3 = Envelope (marque com o número do formato)
   * - **API ID Path**: product.dimensions[].ncdformato
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  ncdformato: prismic.RichTextField;
  /**
   * nVlComprimento field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Comprimento da encomenda (incluindo embalagem), em centímetros
   * - **API ID Path**: product.dimensions[].nvlcomprimento
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  nvlcomprimento: prismic.RichTextField;
  /**
   * nVlAltura field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Altura da encomenda incluindo embalagem, em centímetros. 0 para envelope
   * - **API ID Path**: product.dimensions[].nvlaltura
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  nvlaltura: prismic.RichTextField;
  /**
   * nVlLargura field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Largura da encomenda (incluindo embalagem), em centímetros
   * - **API ID Path**: product.dimensions[].nvllargura
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  nvllargura: prismic.RichTextField;
  /**
   * nCdServico field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: sempre adicione ambos: '04014', '04510'
   * - **API ID Path**: product.dimensions[].ncdservico
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  ncdservico: prismic.RichTextField;
  /**
   * nVlDiametro field in *product → dimensions*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Diâmetro da encomenda (incluindo embalagem), em centímetros
   * - **API ID Path**: product.dimensions[].nvldiametro
   * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
   *
   */
  nvldiametro: prismic.RichTextField;
}